+++ NPCs/FracturedEye.cs
@@ -1,0 +1,55 @@
+using Terraria;
+using Terraria.ID;
+using Terraria.ModLoader;
+
+namespace Harblesnargits_Mod_01.NPCs
+{
+    public class enemy_eye_01 : ModNPC
+    {
+        public override void SetStaticDefaults()
+        {
+            DisplayName.SetDefault("Fractured Eye");
+            Main.npcFrameCount[npc.type] = Main.npcFrameCount[NPCID.DemonEye];
+        }
+
+        public override void SetDefaults()
+        {
+            npc.width = 38;
+            npc.height = 46;
+            npc.damage = 18;
+            npc.defense = 2;
+            npc.lifeMax = 60;
+            npc.HitSound = SoundID.NPCHit1;
+            npc.DeathSound = SoundID.NPCDeath1;
+            npc.value = 75f;
+            npc.knockBackResist = 0.5f;
+            npc.aiStyle = 2;
+            aiType = NPCID.DemonEye;
+            animationType = NPCID.DemonEye;
+            banner = Item.NPCtoBanner(NPCID.DemonEye);
+            bannerItem = Item.BannerToItem(banner);
+        }
+
+        public override float SpawnChance(NPCSpawnInfo spawnInfo)
+        {
+            return !NPC.downedBoss1 ? 0f :
+            SpawnCondition.OverworldNightMonster.Chance * 0.025f;
+        }
+
+        public override void NPCLoot()
+        {
+            if (Main.rand.NextBool(33))
+                Item.NewItem(npc.getRect(), ItemID.Lens, 1);
+            if (Main.rand.NextBool(100))
+                Item.NewItem(npc.getRect(), ItemID.BlackLens, 1);
+        }
+
+        public override void HitEffect(int hitDirection, double damage)
+        {
+            if (npc.life <= 0)
+            {
+                Gore.NewGore(npc.position, npc.velocity, mod.GetGoreSlot("Gores/gore_eye_red"), 1f);
+            }
+        }
+    }
+}
